fix: households provider
