+++ frontend/src/utils/avatarUtils.ts
